Check for wildcard type in adapter factory
Because MDAGsonAdapaterFactory requests collections with a wildcard type
(like `Collection<? extends PackageData>`), we need to add a simple case
to extract the upper bound and match against that if a wildcard is used.

To keep it simple, it only checks for wildcards with 1 upper bound
(so it excludes `? extends PackageData & OtherInterface`).