Convert metadata adapters to use streaming API, add adapter factory
The use of the streaming API increases performance, due to not
requiring Gson to convert the JSON to the tree model (which uses memory)
before passing to the de/serializers.

The use of an adapter factory mimics the MDC adapter factory, and allows
us to pass a Gson instance to the type adapters (for delegation of
deserialization of other metadata objects).